0.0.9
changed the following default settings
- vsync > adaptive (was off)
- particles > decreased (was minimal)
- fullscreen mode > borderless (was fullscreen)
- dynamic lighting > slow (was realtime)
- glint speed & strength > 1.0 (were 0.5 and 0.75 respectively)

added logic to muting the custom background music
added a bug report icon as a universal element. (may tie this to the current screen for help with debugging)
added fade effects to the main menu
customised the pause menu screen. adding a blacklist to not display the universal background
added the same bug reporting button as the one from the main menu and removing the mojang implemented ones.
renamed open to lan to make server (handled by e4mc)

fixed the skin editor button being misaligned

changed settings related to the loading of the window
window now maximises when loading instead of having to wait and the window locking up
added cpu usage to the loading window